FIX: Rubrics and alignment activation should depend on the opened margin.
This commit is the correct implementation of the requirements given in 04524e2edad6d3e4b1bca9932a21958a80ba6360.

The events structure was changed so that the API is not queried a second
time when the margin is opened or closed.

Additional requirement: Rubrics (and "alignment to itself") should be 
activated when the source has rubrics and there is no opened scholium.
This requirement is necessary to reduce layout updates.