FIX: Image references in Markdown may include an alternative text.
The following content complies with the Markdown format:
```
![Me, myself and I](https://fr.gravatar.com/userimage/20324708/0adc3816fe2763cb6b6524bd2638084c.jpeg)
```

Therefore, Hyperglosae should render it with "Me, myself and I" as the
alternative text of the image.

Alternative texts (`alt`) are necessary for people with visual 
impairments as well as for automated tests.

See: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/img#alt