FEATURE: Go to the video excerpt corresponding to a comment (see #74).
With this feature, users will be able to search videos based on comments. In effect, they'll be able to find moments in a video that have already been commented.

So, by clicking on a timecode written in a commentary (glossary document), they'll be able to go directly to the desired excerpt in the video, and the video will automatically jump to the desired timecode.

Co-authored-by: AVCharpentier <antoine_valentin.charpentier@utt.fr>
Co-authored-by: marina-pera <marina.peralta_fonseca@utt.fr>