REFACTORING: Mobile browsing should be kept out of the scope (see #142).
As mobile browsing was developed, especially for self-guided tours, 
it appears that the mobile interface differed more and more from the
desktop one.
Keeping both in the same codebase made the resulting code too complex.
For this reason, the corresponding features (#94, #93 and #99) are now
removed from this application but could be put together in the future 
as a distinct frontend (with the same backend).