REFACTORING: `Page` should be called `Lectern`.
The name "page" could be mistaken for a "Web page" rather than as the 
paper page that can be *read* and *written*, and that contains at the 
same time a *source*, its *references* and its *gloses*.

Moreover the metaphor of a paper page was a bit short to represent that
a glose could be selected (among several ones from different people) to 
be displayed along with the source.

A "lectern" ("lutrin" in French) seems to be a better metaphor.